backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
pymongo
python-dotenv
Werkzeug
cryptography
Pillow
numpy
opencv-python
